Adicionado What's new da próxima versão
- Movido provider do What's new para o diretório da feature;
- Adicionado documentações da versão;
- Adicionado espaço no What's new para aviso.
