+++ src/vscode-whats-new/ui/whats-new.html
@@ -26,9 +26,7 @@
                     ${headerMessage}
                 </p>
                 <blockquote class="header__blurb">
-                    Recomenda-se em ambientes Windows alterar a configuração <code>"advpl.alpha_compile": true</code>
-                    pois as melhorias no Bridge (intermediador do VsCode com AppServer) estão sendo feitas 
-                    somente nessa versão.
+                    <b>Notice:</b> ${noticeMessage}
                 </blockquote>
             </header>
